refactor(renames store-arg store-event, extracts StoreObserver, updates docs)
