+++ libs/core/src/index.ts
@@ -1,5 +1,5 @@
 export * from './lib/store-value';
-export * from './lib/store-arg';
+export * from './lib/store-event';
 export * from './lib/effect';
 export * from './lib/sources';
 export * from './lib/sinks';
